feat(settings): 添加设置模块及相关面板组件
新增设置模块的多个面板组件，包括通用设置、WebDAV配置、备份恢复、分类管理、后端服务和AI配置。每个面板提供相应的功能界面和交互逻辑，支持多语言切换。

- 通用设置面板支持语言切换和版本检查
- WebDAV面板提供配置管理和测试功能
- 备份面板支持数据备份和恢复操作
- 分类面板管理自定义分类和默认分类显示
- 后端面板处理服务器连接和数据同步
- AI面板配置AI服务参数和测试连接
