+++ src/components/settings/index.ts
@@ -1,0 +1,6 @@
+export { AIConfigPanel } from './AIConfigPanel';
+export { WebDAVPanel } from './WebDAVPanel';
+export { BackupPanel } from './BackupPanel';
+export { BackendPanel } from './BackendPanel';
+export { CategoryPanel } from './CategoryPanel';
+export { GeneralPanel } from './GeneralPanel';
